Refactored gem5 code for initializeDatapath() change.
One of the changes in this commit is the TLB initialization.

Previously we insert mapping between trace addresses, simulated virtual
addresses and physical addresses when the program calls
mapArrayToAccelerator(). This call requires arrayBaseAddress is
initialized, which means the graph is constructed and optimizations are
performed.

On the other hand, this refactoring tries to move graph construction
and graph optimization after the user program calls invokeAccelerator(),
which happens after mapArrayToAccelerator(). In this case, Aladdin is
not aware of array's base trace address when mapArrayToAccelerator() is
called since the graph is not constructed.

In fact, we do not need array's base trace address when mapArrayToAccelerator()
is called. All we need is the mapping between array name and it's simulated
virtual address.

In this commit, we switched traceAddrToVirtual mapping to arrayLabelToVirtual
mapping which is initilized when mapArrayToAccelerator() called.
This arrayLabelToVirtual map is independent of whether the graph construction and
transformation are preformed or not.

Change-Id: I080c15bf95fb35eb431c9ae7cc2d16fac37a7366