Support multiple invocations!
The multiple invocation feature is needed when the same function is called
multiple times throughout the program execution.

To support this, we made three major changes:

1) Split the graph building process out of the Datapath contructor:

Originally, the graph is built when the constructor is called. This won't work
for multiple invocations since we want to build the graph when the corresponding
invocation is called. We moved graph building and initialization into the
initializeDatapath() function which is called when the user program calls
invokeAccelerator().

2) Refactor the DDDG to track where it has scanned last time the accelerator is
invoked.

3) Clear the Datapath relavant states at the end of simulation. clearDatapath()
is added to the class and the clear() functions also added to other related
classes.

Change-Id: Iac6015ae999e705ed7dc43d72a87ec553e9b47ef