Fixed deadlock in ls queue when there are many cocurrent accesses.
Two checking logics fixed:

1. After TLB miss returns, it sets the node status to Ready so it can reissue
the tlb translation request. In this case, the address is in the queue already.
If we check whether the queue is full first before we check whether the queue
contains the address, it will not handle (enqueue in this case) the node at all,
which causes deadlock (the queue is always full and never gets cleared.)

2. When the address is translated, in the issueCacheRequest(), we first check
whether the queue.can_issue() or not. At this point, the addr is already in the
queue. The queue.can_issue() logic somehow inclues whether the queue.is_full()
or not, which has nother with whether the queue can issue or not. In a situation
where the queue is full but has enough bandwidht, it will never make progress.

Also apply the same addr mask when cache returns.

Change-Id: I171e20aef9936ef1e9ffff500d128bb3459bf3ab