Automated integration test framework for Jenkins.
This framework can validate performance statistics and whether the test
crashed or not. All output is stored in a temp dir and cleaned at the
conclusion of each test if PRESERVE_RUN_DIR is defined in the user
environment.

This framework will check the simulator return code, so some exit()
statements have been modified to return 1 on error instead of 0.

In addition, each test section name has been changed to match the name
of the test itself, both for clarify and for being able to infer things
in the testing scripts.

BUG=ALADDIN-30
Change-Id: I907ff1faa7755cdeb0feb8d296de3dc08626cab2