Speed up unit test build times.
Because Catch is a header-only library, if we #define CATCH_CONFIG_MAIN at the
top of each unit test, we end up compiling Catch once per unit test, which is
wasteful. Instead, we will compile it just once in its own translation unit and
then link all the unit test objects against it.

A quick test on rb22 showed a speedup from 2m22s to 1m15s. This isn't such a
problem on our servers becase we can do parallel builds, but on Jenkins where
we've only provisioned one core, it will help a lot.

Change-Id: I8adf9123f85751ff7049ab70819a77e52ba5a0a8