Add feature to dump or reset gem5 stats during CPU simulation.
This is implemented using the ioctl syscall, but with a special request
code that cannot be assigned to an ordinary accelerator. When the dump
function is called, a description string can be passed to this. This
will cause the description to be printed inline with the stats file and
should be used to attach a helpful description of what the stat dump
represents (such as "completed reading file") since the last call to
resetGem5Stats().

The difference between this commit and the previous commit on dumping
gem5 stats is that the previous one only supports dumping stats after
accelerator invocations, whereas this one is for arbitrary CPU
simulations.

This requires support from the simulator too, obviously.

Change-Id: I0ea62cc88bd6c2dc47ec9272a7fc99e2c2a6e74b