Support storing floating point values in Aladdin simulation.
Before we only support integer values, where all the values stored from Aladdin
are converted to long long int. Here we store all the values in double but keep
track of its original type. When we do a store in Aladdin simulation, we check
its original type and convert the double to its orignal type so that the value
stored in memory match its original format.

Change-Id: Ife09b08be1a4a4375cb5a42707974b2699f6490a