Remove all global using namespace std declarations.
This was partly done to help in resolving the boost graphviz compilation
failures that had required changing the Boost headers themselves (which is
extremely evil). A namespace collision on the function get() was what I
originally suspected to be the problem. It turns out that wasn't the problem,
but this is better anyways.

In some cases, using this at the function-level is cleaner, particularly in
functions that just print to std::cout.

BUG=ALADDIN-43

Change-Id: Ic6ab7804cfe73f76c4dd3c62dbb114290ad37efa