Eliminate ioctl request code checking entirely.
Instead of checking if the code is within a valid range AND it does not
belong to a special value, just check if the file descriptor for the
ioctl call is ALADDIN_FD or not. We always forget to update the valid
range anyways (ideally this would be done by default), and even that
doesn't really guarantee that the code is really valid (i.e. it IS
assigned to an accelerator that EXISTS and is REGISTERED with the
system).

Change-Id: I0ce7e3e5e78f66bbee2f456bcb848868aab4b5c4