Fixed the dmaStore hack that only issues dmaStore in the end.
Finally tracked down the issue with dmaStore. We had an ugly hack to force
dmaStores to be issued in the very end (total - 5 nodes) because we observed that
dmaStores could be issued earlier than when they are supposed to be. The reason
for this is how we handle consecutive dma operations. The observed bug happened
when there are multple dmaStores.

Originally in order to avoid blocking multiple consecutive dmaLoads, since
multiple dmaLoads can be issued in parallel, we add special conditions (e.g., if
both prev_branch and current node are dma operations) to avoid adding edges
between dma operations. One issue here is that by not adding edges between two
consecutive dma operations, it is possible that we make the consecutive dma
operations not dependent on any nodes. This is fine for dmaLoads, as they are
indeed usually not dependent on anything since they are at the beginning of a
trace. But this will cause issues for dmaStores: if a dmaStore is not dependent
on any nodes, it will be scheduled at the very beginning.

To fix this, we handle the consecutive case more carefully at two places:

1) When we see two consecutive DMA operations, we still do not add a control
edge between them, as they can be issued concurrently. But we keep track of the
last DMA node in the group, so that the following instructions can not be issued
until all the DMA nodes in the group are completed.

2) After that, we do not treat DMA nodes as branch/boundary nodes anymore. We
just treat them as non-branch nodes that are guarded by branch/boundary nodes.
They cannot be issued until all the previous instructions are completed.

We are still using control dependence to schedule dma operations, as we are not
checking the addresses of dma operations and the memory loads/stores in the body
of the program. We assume when programmers write program, the ordering of DMA
operations and the program body is already considered.

Also modified the number of expected nodes in the unit test. The reason for the
change is because we also add vertex when we instantiate a new exec_node. Before
vertexes in the boost graph is only added when we call add_edge(). In this case,
isolated nodes from DDDG constructions will not be added in the graph. However,
for DMA stores at the end of the trace, although it's not connected with anyone
in DDDG but it cannot be eliminated in the graph. Although the unit test trace
doesn't have DMA instructions, the isolated nodes in the end, like return op,
are included in the graph, which causes the increase in the number of nodes.

BUG=ALADDIN-23

Change-Id: Iaa6ff6da45a81dd29064bdaab7742acdf23374bd