A HACK FOR AES SBOX.
sbox is a constant array inside an aes accelerator. In the ready bit simulation,
it should be set by default, though currently we don't have a mechanism to
differentiate a const array to a regular array, which is definitely doable.
Since it only happens to aes for now, we use this temporary hack to get around
with this issue for now, but we should come back to fix it later.

Change-Id: I909fcbdcf635cd9d1a9997629b8e4e183b88e35c