Added a mapping between register IDs in caller and callee functions.
To fix a global-local variable name mapping issue that comes up in
sort-radix, where we have:

if (flag == 0) {
  update(b, bucket, a);
} else {
  update(a, bucket, b);
}

Both a and b are arrays. update is declared as this:

void update(int b[N], int bucket[BUCKETSIZE], int a[N]);

In this case, the accesses to b in update() could go either array a or b,
depends on whether the flag is set.

We already have mechanisms to pass dependence across functions with the forward
flag, where we can trace back to when a register was initially defined in the
caller function. For example, if we have

r1 = c[5];
r2 = d[10];
update(r1, bucket, r2);

Then we know accesses to b in update() maps to r1, then mapped to c, while a
mapped to r2 which is mapped to d.

However, in this case, both registers a and b are defined (allocated) outside
the kernel. Even with the forward case, there is no instruction in the caller
function that defines a or b when calling update() function. To handle this
case, we add a table to keep track of the mapping between parameters in each
dynamic call function, in which we have:

(the first time update() was called, flag == 0)
ss_sort-0-b --> update-0-b,
ss_sort-0-a --> update-0-a,
(the second time update() was called, flag != 0)
ss_sort-0-a --> update-1-b,
ss_sort-0-b --> update-1-a,

In this way, we can trace back to which array it's "really" accessing. Also
added a unit test for this.

Change-Id: I918c9e807206d9f7b47fc2d66d3b50f5542fccca