dma: Order DMA nodes by memory dependences, not control dependences.
In the past, DMA nodes were ordered by artificially introducing control
dependencies in the loop unrolling phase, based on some common
programming patterns (like only adding a control edge on the last of a
group of DMA nodes). Now, DMA loads and stores are treated like normal
loads and stores, so they use memory dependences that are added during
DDDG construction in an identical fashion to normal loads and stores.

Because DMA loads (which are stores from the accelerator's perspective)
are variable-latency operations, we must enforce WAW dependencies on
normal stores (because repeated stores to the same address with no
intervening loads are typically optimized away).

This commit also fixes a bug in loop unrolling such that non-branch
nodes are never confused as loop boundaries (which was discovered while
debugging extraneous control edges).

Change-Id: Idd106867e545fc8b1f9581ea888b74405f3f3317