dddg: Extend the string refactor to the datapaths.
(Almost) every string is now stored in the SourceManager, and ExecNodes
now should only hold ids or references to objects in the SourceManager.
The final way to access SourceManager is:

  T& ref = srcManager.insert<T>(object_name)
  const T& ref = srcManager.get<T>(object_name)
  const T& ref = srcManager.get<T>(object_id)

where T is the type of SourceEntity.

I've left array labels as strings for now because they are extensively
used during scheduling, and replacing them with ids might incur a
performance hit. Basic block ids don't seem to be used at all - those
will be removed in a later commit.

Change-Id: I4101f2d3338d879a81cf6ce9960b90c96d795801