Add basic performance tests.
Run this standalone to measure wall clock time for different traces, or
run this under valgrind to measure memory usage (and possibly locate
memory leaks).

Build with `make perf`, clean with `make clean-perf` (this test is not a
unit test that checks for correctness, so it does not use the Catch
framework).

Change-Id: Id8cda4e18a6cc77519a85f76a35f2aa456eb3bbd