dddg: Fix bug with renamed arrays as function arguments.
If an array is renamed in a child function's argument list, during DDDG
construction, we attempt to update the base address of the renamed array,
rather than that of the real array. Since the renamed argument would not appear
in the configuration file, nothing actually gets changed. Furthermore, if the
real array is never updated under its own name at a later stage, then it will
not have a base address.

Fix: when setting a base address on an array for a GEP node, always backtrace
the array name to its original name in the top level function, and update the
base address for that array instead.

Added integration test for this corner case as well.

Example code:

  void child_func(TYPE* array, size_t idx) {
    /* 'array' could mean either array0 or array1. */
  }

  void kernel(TYPE* array0, TYPE* array1, size_t num_vals) {
    loop: for (int i = 0; i < num_vals; i++) {
      if (/* condition */)
	child_func(array0, i);
      else
	child_func(array1, i);
    }
    /* array0 and array1 are never modified in this function. */
  }

BUG=ALADDIN-74

Change-Id: I9bf80c2784df0d2127575c6bc422b6aa2625865c