base: Change OutEdgeList from listS to setS for Graph.
OutEdgeList goes from listS to setS because this greatly accelerates
clear_vertex() on large, densely connected graphs (changes a linear scan
to find each edge in the edge list to a logarithmic scan). For example,
when unrolling loops with a high partitioning factor, we clear vertices
on a large number of branch nodes. This can accelerate clear_vertex from
300-400 vertices/sec to 1M vertices per sec.

This change also revealed a bug in DDDG generation: we can have multiple
identical edges between two nodes, but identical edges are disallowed by
setS.

The one drawback of going to setS is the increased memory usage per
vertex, but the speedup obtained makes this cost well worth it. One
future optimization to counter this is to remove isolated vertices (and
ExecNodes) from the graph entirely, but this could be quite tricky.

Change-Id: Id35f2d91f0ade17d1596e0d09bf56b42a2c69dea