base: Change VertexProperty from vertex_index_t to a custom property.
The custom property is called vertex_node_id_t, indicating that it
stores the node id of the vertex.  vertex_index is an built-in property
which uniquely identifies the vertex, so it has to start from 0 and
count up, but node_id can be anything - there could be missing nodes, or
we could start from a nonzero index. The custom property allows us to
decouple these two properties.

Also, it appears that boost_typedefs.h was reintroduced at some point,
but it was already renamed to typedefs.h, so boost_typedefs.h is getting
deleted.

Change-Id: I98da688cc7e2816b7b372abc5e5980de6bcbd9c3