HybridDatapath: improve datapath naming and CommMonitor setup.
Datapaths are now named "ACCELNAME_datapath", where ACCELNAME is the
name of the section in gem5.cfg describing this datapath. It is no
longer named as "datapathN" with N = accelerator id. The new naming
improves readability of the debugging traces.

CommMonitors are now attached as children of the HybridDatapath object,
rather than children of the top level System. This just makes more sense
(and also makes the names of the statistics consistent).

Change-Id: I7400705c793fb6ae327c033f7444e0c3c940f618