debugger: Add a debugger for Aladdin.
Using gdb to debug issues with the DDDG is incredibly difficult and
tedious (not all functions are visible to or callable from gdb), and
rendering the entire DDDG is basically impossible for any graph
visualization software I know of. So instead, I wrote an interactive
debugger that will print information about nodes and the graph. It works
as a drop-in replacement for Aladdin, but it will interrupt the flow of
Aladdin at specific places to allow the user to query for information.

Change-Id: I2c5cb44d4098ffad6e35395277d6d85fe8e70d73