core: Revamp the loop unrolling algorithm.
Take advantage of loop depth information provided by LLVM-Tracer to keep
track of loop hierarchy, so we can know when we exit an inner loop to
its parent loop. This thus informs us when to reset the iteration
counter for that loop (which could not happen before).

This fixes various unrolling bugs which stem from not resetting the
counter, as without loop hierarchy information it is impossible to know
when to reset.

This requires LLVM-Tracer commit 39c051, but it is backwards compatible
with previous versions of the tracer, so without loop depth information,
the old behavior is preserved.

Change-Id: Ief999e1c91aeb2beac3461a24d1a1e79f73f9ae4