common: Properly enforce hard dependences across function calls.
A function cannot be called until every preceding node has executed, and
a called function must execute to completion before any node in the
parent node can execute. The only exceptions are DMA nodes.

This may be something we want to revisit or make user-controllable (with
something like fences), but without this, we could have multiple
instances of a function executing at once, or function invocations
becoming "pipelined", when we don't expect them to.

Change-Id: I06f2a33efdf5f30cba17c4d85ceca78c30cdc320