gem5: Implement a basic model for ARM ACP.
The ACP port is connected to the L2 cache crossbar. It issues normal
read requests, but writes require any modified cache lines in L1 caches
to be cleaned and evicted to L2 first. This implementation
handles this requirement by issuing an exclusive read request,
which directly transfers ownership to the ACP port without first
writing back to L2.

There is also no need to issue multiple ownership requests to the same
cache line, but this implementation also does not handle that yet
either.

Tested with a basic integration test, but more testing is needed.

Change-Id: I6c0962d6c40b0de311f04d4dde74e44c4502dbad