gem5: Improve the AladdinTLB invisible translation API.
Instead of modifying the MemAccess objects directly in the translation
functions, return a struct that contains the virtual and physical
simulation addresses, and let the caller decide what to do with it. This
is better encapsulation of functionality.

This is setting up the next commit, which will get rid of the physical
address field in MemAccess entirely.

Change-Id: I08d372e9416bdd2d0e0ba077174bc9f5f5133eea