gem5: New ready-bit programming model.
When ready_mode is enabled, use the new setReadyBits() function to
programmatically change the state of ready bits on the scratchpads. No
memory or DMA accesses to the address range specified will execute until
this update has happened. DMA loads will still update ready bits to 1.

When ready_mode is disabled, setReadyBits() nodes will be removed from
the graph so that they have no effect, but a warning will be printed to
remind the user of this fact.

There is currently no way to overlap computation with DMA stores. This
requires the compute operations to change the state of ready bits to
trigger the DMA store to proceed, which is nontrivial because it would
require the DMA store node to be broken up into many separate nodes that
could be independently scheduled.

TESTED=integration

Change-Id: I23840cbf48e7c0f7ffea740bd7d29e9ead0f40e0